Finish db initialization; Add syncing of file tree items
