+++ backend/internal/database/checkIfFileIsInBatchByPath.go
@@ -2,7 +2,7 @@ package database
 
 import "backend/pkg/models"
 
-func CheckIfFileIsInBatchByPath(path string, batch []models.FileTreeItem) bool {
+func checkIfFileIsInBatchByPath(path string, batch []models.FileTreeItem) bool {
 	doesFileWithPathExist := false
 	for _, referenceFile := range batch {
 		if referenceFile.Path == path {
