--- backend/internal/database/getSQLFileContent_test.go
@@ -1,1 +1,0 @@
-package database
