+++ backend/internal/database/syncFileTree.go
@@ -1,0 +1,153 @@
+package database
+
+import (
+	"backend/pkg/constants"
+	"backend/pkg/enums/fileType"
+	"backend/pkg/models"
+	imageConverterUtilities "backend/pkg/services/imageConverter/utilities"
+	"context"
+	"log"
+	"strings"
+
+	"gorm.io/gorm"
+)
+
+func syncFileTree(databaseConnection *gorm.DB, filesFromDisk []models.FileTreeItem, filesFromDatabase []models.FileTreeItem, ctx context.Context) error {
+	filesToDelete := getDistinctFiles(filesFromDatabase, filesFromDisk)
+	filesToCreate := getDistinctFiles(filesFromDisk, filesFromDatabase)
+	deleteError := deleteFileTreeItemsFromDb(databaseConnection, ctx, filesToDelete)
+	if deleteError != nil {
+		return deleteError
+	}
+
+	insertError := insertFileTreeItemsIntoDb(databaseConnection, ctx, filesToCreate)
+	if insertError != nil {
+		return insertError
+	}
+
+	return nil
+}
+
+func doesFileTreeItemWithFileIdExitsInDb(databaseConnection *gorm.DB, fileId string) bool {
+	count := int64(0)
+	databaseConnection.Model(&models.FileTreeItem{}).Where("file_id = ?", fileId).Count(&count)
+	return count > 0
+}
+
+func deleteFileTreeItemsFromDb(databaseConnection *gorm.DB, ctx context.Context, filesToDelete []models.FileTreeItem) error {
+	if len(filesToDelete) == 0 {
+		return nil
+	}
+
+	for _, fileToDelete := range filesToDelete {
+		_, deleteError := gorm.G[models.FileTreeItem](databaseConnection).Where("id = ?", fileToDelete.ID).Delete(ctx)
+		if deleteError != nil {
+			log.Printf("Error while deleting file: %v\n", fileToDelete.Path)
+			return deleteError
+		}
+
+		log.Printf("Successfully deleted file: %v\n", fileToDelete.Path)
+
+		shouldTryCascadingDeletionOfAssociatedAudioFile := fileToDelete.AssociatedAudioFileId != nil
+		if shouldTryCascadingDeletionOfAssociatedAudioFile {
+			doesAudioFileExits := doesFileTreeItemWithFileIdExitsInDb(databaseConnection, *fileToDelete.AssociatedAudioFileId)
+			if doesAudioFileExits {
+				_, deleteError = gorm.G[models.FileTreeItem](databaseConnection).Where("file_id = ?", fileToDelete.AssociatedAudioFileId).Delete(ctx)
+				if deleteError != nil {
+					log.Printf("Error while deleting audio file with fileId='%v' of subtitle file with path='%v'\n", fileToDelete.AssociatedAudioFileId, fileToDelete.Path)
+					return deleteError
+				}
+
+				log.Printf("Successfully deleted audio file with fileId='%v' of subtitle file with path='%v'\n", fileToDelete.AssociatedAudioFileId, fileToDelete.Path)
+			}
+		}
+
+		shouldTryCascadingDeletionOfAssociatedLowQualityImage := fileToDelete.LowQualityImageId != nil
+		if shouldTryCascadingDeletionOfAssociatedLowQualityImage {
+			doesLowQualityImageExist := doesFileTreeItemWithFileIdExitsInDb(databaseConnection, *fileToDelete.LowQualityImageId)
+			if doesLowQualityImageExist {
+				_, deleteError = gorm.G[models.FileTreeItem](databaseConnection).Where("file_id = ?", fileToDelete.LowQualityImageId).Delete(ctx)
+				if deleteError != nil {
+					log.Printf("Error while deleting low quality image with fileId='%v' of item with path='%v'\n", fileToDelete.LowQualityImageId, fileToDelete.Path)
+					return deleteError
+				}
+
+				log.Printf("Successfully deleted low quality image with fileId='%v' of item with path='%v'\n", fileToDelete.LowQualityImageId, fileToDelete.Path)
+			}
+		}
+	}
+
+	log.Printf("Deleted '%v' files\n", len(filesToDelete))
+	return nil
+}
+
+func insertFileTreeItemsIntoDb(databaseConnection *gorm.DB, ctx context.Context, filesToAddInput []models.FileTreeItem) error {
+	if len(filesToAddInput) == 0 {
+		return nil
+	}
+
+	filesToAdd := make([]models.FileTreeItem, len(filesToAddInput))
+
+	/**
+	There are two cases files and their associated files can be added:
+		1. On the same batch: The associated file is added in the same batch from the file system as the file, so it is part of the input => do nothing
+		2. On a postposed batch: The associated file is not part of the same batch from the file system but could be already added to the database => try to get ids from the database
+
+	Only these types of files have this problem:
+		- subtitle
+		- image
+	*/
+	for i, initialFileToAdd := range filesToAddInput {
+		if initialFileToAdd.Type == fileType.IMAGE {
+			isLowQuality := imageConverterUtilities.IsLowQualityImagePath(initialFileToAdd.Path)
+			if isLowQuality {
+				pathWithoutLowQualitySuffix := imageConverterUtilities.RemoveLowQualitySuffixFromImageName(initialFileToAdd.Path)
+				isSameBatchInsert := checkIfFileIsInBatchByPath(pathWithoutLowQualitySuffix, filesToAddInput)
+
+				if !isSameBatchInsert {
+					matchingImage, tryGetMatchingImageByPathError := tryGetFileTreeItemByPath(databaseConnection, ctx, pathWithoutLowQualitySuffix)
+					if tryGetMatchingImageByPathError != nil {
+						log.Printf("Warning: No matching audio file with path='%v' found. error='%v'\n", pathWithoutLowQualitySuffix, tryGetMatchingImageByPathError.Error())
+					}
+
+					if matchingImage != nil && matchingImage.ID > 0 {
+						initialFileToAdd.LowQualityImageId = &matchingImage.FileId
+					}
+				}
+
+			}
+		}
+
+		if initialFileToAdd.Type == fileType.SUBTITLE {
+			pathOfPossibleMatchingAudioFile := strings.TrimSuffix(initialFileToAdd.Path, constants.SubtitleExtension)
+			isSameBatchInsert := checkIfFileIsInBatchByPath(pathOfPossibleMatchingAudioFile, filesToAddInput)
+
+			if !isSameBatchInsert {
+				matchingAudioFile, tryGetMatchingAudioFileByPathError := tryGetFileTreeItemByPath(databaseConnection, ctx, pathOfPossibleMatchingAudioFile)
+				if tryGetMatchingAudioFileByPathError != nil {
+					log.Printf("Warning: No matching audio file with path='%v' found. error='%v'\n", pathOfPossibleMatchingAudioFile, tryGetMatchingAudioFileByPathError.Error())
+				}
+
+				if matchingAudioFile != nil && matchingAudioFile.ID > 0 {
+					initialFileToAdd.AssociatedAudioFileId = &matchingAudioFile.FileId
+				}
+			}
+		}
+
+		filesToAdd[i] = initialFileToAdd
+	}
+
+	result := gorm.WithResult()
+	createInBatchError := gorm.G[models.FileTreeItem](databaseConnection, result).CreateInBatches(ctx, &filesToAdd, len(filesToAdd))
+	if createInBatchError != nil {
+		log.Println("Error creating files")
+		return createInBatchError
+	}
+
+	return nil
+}
+
+func tryGetFileTreeItemByPath(databaseConnection *gorm.DB, ctx context.Context, path string) (*models.FileTreeItem, error) {
+	matchingFile, err := gorm.G[*models.FileTreeItem](databaseConnection).Where("path = ?", path).First(ctx)
+	return matchingFile, err
+}
